drivers: spi: Add Target mode support for DesignWare SPI controller
Added support for configuring the DesignWare SPI controller
to operate in target mode.

This patch enables the DesignWare SPI controller to function
in target mode. It updates the driver to configure the necessary
registers and handle target-specific SPI transactions.
This is required for systems where the SPI controller must act
as a peripheral device rather than a master.

Signed-off-by: Aravind Krishnan <aravind.krishnan@alifsemi.com>