dt-bindings: sdhci-reset: Add Alif SDHCI-RESET device tree bindings
This patch introduces a new device tree binding for the Alif Semiconductor
SDHCI (Secure Digital Host Controller Interface) reset controller. This
binding describes the reset controller properties required for integrating
SDHCI controllers on Alif platforms.

It defines the compatible string and documents the expected reset control
behavior in the context of Alif’s SoC architecture. This is required for
correct initialization and control of the SD card interface.

Signed-off-by: Pankaj Pandey <pankaj.pandey@alifsemi.com>