arm: dts: alif: Remove unused DAC122_STATUS macro from DCT header
The DAC122_STATUS macro is defined in the DCT (Device Configuration Tool)
header file but corresponds to a DAC instance (DAC122) that is not
present or supported in the current hardware or device tree.

This unused define adds unnecessary clutter and may mislead developers
into thinking DAC122 is available. Since there is no corresponding
device node or driver support for DAC122, this macro should be removed
to keep the configuration accurate and minimal.

Signed-off-by: Akshay-AlifSemi <raghavendra.akshay@alifsemi.com>