arm: dts: alif: Add no-1-8-v for SDHCI on Eagle Devkit
The 16 GB SanDisk MicroSDHC card fails to be detected
on the Alif Ensemble-E8 platform due to the SDHCI
controller attempting to switch to 1.8V signaling
during initialization. However, the hardware only
supports 3.3V I/O signaling and lacks a controllable
VQMMC regulator.

Adding the 'no-1-8-v' property disables 1.8V voltage
switching, forcing the controller to remain in 3.3V
mode, which allows the card to be properly detected
and initialized.

This change is required for reliable operation with
standard SDHC cards on this hardware.

Signed-off-by: Akshay-AlifSemi <raghavendra.akshay@alifsemi.com>