arm: dts: alif: Add shared cmp_base to ADC node
The ADC12 (0/1/2) and ADC24 instances share the
same comparator block at 0x49023000. To allow
the driver to map this region, include it as a
second memory resource ("cmp_base") to ADC
device tree node.

This enables the driver to access the shared
block regardless of which ADC instance is
enabled (ADC12 group or ADC24).

Signed-off-by: Akshay-AlifSemi <raghavendra.akshay@alifsemi.com>